requirements.txt
flask
gensim
sastrawi
python-docx 
pdfplumber
nltk
werkzeug
pymysql
psycopg2
PyPDF2
spacy
reportlab